[Feat] Add getMapboxRef prop (#372)
* [Feat] Add getMapboxRef prop

The getMapboxRef prop can be set to a function which will gain access to KeplerGl's MapContainer's map (a react-map-gl InteractiveMap instance). In other words, it gives KeplerGl's parent component a reference to the inner Mapbox map.

* Forward getMapboxRef, support multiple maps (toggleSplitMap)

This simplifies the handling of getMapboxRef so that KeplerGL forwards your callback function to MapContainer unchanged. It also adds support for multiple maps. Use 'index' (0 or 1) to identify which map triggered the callback. If 'mapbox' is null, then that map has been removed by toggleSplitMap.

* Add getMapboxRef prop to README